+++ layers/CMSSchema/packages/custompost-tag-mutations/src/Hooks/MutationResolverHookSet.php
@@ -10,6 +10,7 @@ use PoPCMSSchema\CustomPostTagMutations\FeedbackItemProviders\MutationErrorFeedb
 use PoPCMSSchema\CustomPostTagMutations\Hooks\AbstractMutationResolverHookSet;
 use PoPCMSSchema\Tags\TypeAPIs\QueryableTagTypeAPIInterface;
 use PoPCMSSchema\Tags\TypeAPIs\TagTypeAPIInterface;
+use PoPCMSSchema\Taxonomies\TypeAPIs\TaxonomyTermTypeAPIInterface;
 use PoP\ComponentModel\Feedback\FeedbackItemResolution;
 use PoP\ComponentModel\Feedback\ObjectTypeFieldResolutionFeedback;
 use PoP\ComponentModel\Feedback\ObjectTypeFieldResolutionFeedbackStore;
@@ -18,6 +19,7 @@ use PoP\ComponentModel\QueryResolution\FieldDataAccessorInterface;
 class MutationResolverHookSet extends AbstractMutationResolverHookSet
 {
     private ?QueryableTagTypeAPIInterface $queryableTagTypeAPI = null;
+    private ?TaxonomyTermTypeAPIInterface $taxonomyTermTypeAPI = null;
 
     final public function setQueryableTagTypeAPI(QueryableTagTypeAPIInterface $queryableTagTypeAPI): void
     {
@@ -32,10 +34,23 @@ class MutationResolverHookSet extends AbstractMutationResolverHookSet
         }
         return $this->queryableTagTypeAPI;
     }
+    final public function setTaxonomyTermTypeAPI(TaxonomyTermTypeAPIInterface $taxonomyTermTypeAPI): void
+    {
+        $this->taxonomyTermTypeAPI = $taxonomyTermTypeAPI;
+    }
+    final protected function getTaxonomyTermTypeAPI(): TaxonomyTermTypeAPIInterface
+    {
+        if ($this->taxonomyTermTypeAPI === null) {
+            /** @var TaxonomyTermTypeAPIInterface */
+            $taxonomyTermTypeAPI = $this->instanceManager->getInstance(TaxonomyTermTypeAPIInterface::class);
+            $this->taxonomyTermTypeAPI = $taxonomyTermTypeAPI;
+        }
+        return $this->taxonomyTermTypeAPI;
+    }
 
     /**
      * Retrieve the taxonomy passed via the `taxonomy` input.
-     * If that's not possible (eg: on `createCustomPost.setTags`),
+     * If that's not possible (eg: on `createCustomPost:input.tagsBy`),
      * then retrieve it from queried object's CPT.
      */
     protected function getTagTaxonomyName(
